Made MaybeValue.
The idea of MaybeValue is to store integers and lowed the work at Operation. Tests has been added to instances of MaybeValue. LiftA2 was changed to get a IntBinaryOperation (which translates in (Int -> Int -> Int) in a Haskell syntax). It was found when writing the tests that the design of using operation it's quite verbose and was not really Open-Closed. The implementation arose doubts about which pattern to use at Operation, but this will be discussed in the next commit.