Operation changed from a Template method pattern to a Strategy pattern
First, the design using liftA2 idiom was easier, and the change from a Template Method Pattern to a Strategy pattern is not that much of a design problem. The problem was "overengineering" the part of a division with 0. Having an Either idiom would be probably the best solution to this problem, but it seemed a bit much overwork for something that may never happen. Instead, the Exception approach is a bit less "clean" in the sense that when an ArithmeticException occurs you won't know, as it gets crunched in the Nothing of MaybeValue, but the overwork is small and it accepts the change.