I wanted more tests, and found a bug in Operation.
Operation had implemented an equals without using the function parameter. This is quite troubling as Sum and Mult stand the same, but what's even more troublesome is that comparing functions is undecidable, so the equals was impossible to get in the first place. Well, we now have an equals in reference that could compares references when comparing Operations. But it's still "needed" from tests, and it shouldn't be a problem, as the Façade hides the equals functionality returning always Either.